Some improvements on types and constructors

+++ CocoaHTTPServer/Core/HTTPResponse.h
@@ -7,7 +7,7 @@
  * Returns the length of the data in bytes.
  * If you don't know the length in advance, implement the isChunked method and have it return YES.
 **/
-- (UInt64)contentLength;
+- (NSUInteger)contentLength;
 
 /**
  * The HTTP server supports range requests in order to allow things like
